+++ CHANGELOG.md
@@ -11,6 +11,7 @@ Date: 2022-11
 * Bug fix for parsing raw time limit from sacct (issue #58).
 * Bug fix for Job class when start time is not set (issue #62).
 * Bug fixes for handling jobs that are cancelled whilst pending (issues #63 and #65).
+* Bug fix for handling scontrol values that have spaces in them (issue #67).
 * Added new "never-ran" e-mail template (issue #63).
 * Added the ability for users to configure the e-mail regular expression in slurm-mail.conf (issue #54).
 * Added support for optionally running commands after submitting jobs when running integration tests.